refactor(controller): move clients under controllers
