--- src/main/kotlin/sc/gui/AbstractGuiClient.kt
@@ -1,5 +1,0 @@
-package sc.gui
-
-import sc.plugin2021.AbstractClient
-
-abstract class AbstractGuiClient(host: String, port: Int): ClientInterface, AbstractClient(host, port)
