+++ src/main/kotlin/sc/gui/controller/client/AbstractGuiClient.kt
@@ -1,0 +1,5 @@
+package sc.gui.controller.client
+
+import sc.plugin2021.AbstractClient
+
+abstract class AbstractGuiClient(host: String, port: Int): ClientInterface, AbstractClient(host, port)
