+++ src/main/kotlin/sc/gui/controller/client/ClientInterface.kt
@@ -1,4 +1,4 @@
-package sc.gui
+package sc.gui.controller.client
 
 interface ClientInterface {
     fun joinPreparedGame(reservation: String)
